PW11 - Добавить Контекс текущего пользователя

+++ src/utils/api.js
@@ -78,22 +78,23 @@ class Api {
     }).then(this._checkResponse);
   }
 
-  addLikeToCard(cardId) {
-    return fetch(this._baseUrl + `/cards/${cardId}/likes`, {
-      method: "PUT",
-      headers: {
-        authorization: this._authorization,
-      },
-    }).then(this._checkResponse);
-  }
-
-  removeLikeFromCard(cardId) {
-    return fetch(this._baseUrl + `/cards/${cardId}/likes`, {
-      method: "DELETE",
-      headers: {
-        authorization: this._authorization,
-      },
-    }).then(this._checkResponse);
+  changeLikeCardStatus(cardId, isLiked) {
+    if(isLiked) {
+      return fetch(this._baseUrl + `/cards/${cardId}/likes`, {
+        method: "DELETE",
+        headers: {
+          authorization: this._authorization,
+        },
+      }).then(this._checkResponse);
+    }
+    if(!isLiked) {
+      return fetch(this._baseUrl + `/cards/${cardId}/likes`, {
+        method: "PUT",
+        headers: {
+          authorization: this._authorization,
+        },
+      }).then(this._checkResponse);
+    }
   }
 }
